Submitting Capstone Project 1: Project Proposal
Submitting my proposal for the first Capstone project in Springboard's Data Science Career Track. This is the initial copy and may potentially be updated at a later date. It gives a high-level overview of what I am doing and a general outline of what the next steps are, which are also subject to change.